+++ rigid_force.h
@@ -3,6 +3,6 @@
 
 # include "defs.h"
 
-void forceUpdate( vector<SubData>& particle,  double *p_energy , int* combine_now , int combine[][4], int* step ,const int NrParticles, const double Lx,  const double Ly,  const double Lz );
+void forceUpdate( vector<SubData>& particle,  double *p_energy , int* combine_now , int combine[][4], long long int* step ,const int NrParticles, const double Lx,  const double Ly,  const double Lz );
 
 #endif
